refactor: remove unusued fetch-lib command
This removes the unused pre-craft-application `fetch-lib` command.
